requirements.txt
discord.py>=2.5.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
Pillow>=10.4.0,<12.0.0
